refactor(app): extract block management callbacks into useBlockManagement (cfo)
Creates src/hooks/useBlockManagement.ts (~240 lines) to own all block/group
CRUD operations: updateBlock, updateGroup, updateBlockPositions,
updateGroupPositions, addBlock, handleCreateBlockConfirm,
handleCreateBlockFromCanvas, deleteBlock, deleteBlockWithFile,
getSelectedFolderForNewBlock.

Extracts buildNewBlockContent to src/lib/blockContent.ts as a pure function
(was a no-dep useCallback in App.tsx).

App.tsx: 2686 → 2464 lines (−222). Target <2,500 lines reached. 686 tests pass.
